requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
coqui-tts==0.27.5
transformers>=4.57,<5.0
torch>=2.1
torchaudio>=2.1
librosa>=0.10.2
soundfile==0.12.1
pydub==0.25.1
numpy>=1.26
